change to float in models

+++ base/models.py
@@ -20,7 +20,7 @@ class user_goals(models.Model):
     experience = models.CharField(max_length=25, choices=experience_choices, default='beginner')
     tdee = models.IntegerField(default=1000)
     pRatio = models.IntegerField(default=50)
-    weekly_target = models.IntegerField(default=0, validators=[
+    weekly_target = models.FloatField(default=0, validators=[
             MaxValueValidator(100),
             MinValueValidator(-100)
         ])
